reorganise commands and move to use new lagoon package
