+++ cmd/delete.go
@@ -5,7 +5,6 @@ import (
 	"os"
 
 	"github.com/amazeeio/lagoon-cli/graphql"
-
 	"github.com/spf13/cobra"
 )
 
@@ -23,4 +22,24 @@ var deleteCmd = &cobra.Command{
 			}
 		}
 	},
+}
+
+var deleteVariableCmd = &cobra.Command{
+	Use:   "variable",
+	Short: "Delete variables from environments or projects",
+	PersistentPreRun: func(cmd *cobra.Command, args []string) {
+		// get a new token if the current one is invalid
+		valid := graphql.VerifyTokenExpiry()
+		if valid == false {
+			loginErr := loginToken()
+			if loginErr != nil {
+				fmt.Println("Unable to refresh token, you may need to run `lagoon login` first")
+				os.Exit(1)
+			}
+		}
+	},
+}
+
+func init() {
+	deleteCmd.AddCommand(deleteVariableCmd)
 }
