+++ cmd/project.go
@@ -1,14 +1,14 @@
 package cmd
 
 import (
+	"encoding/json"
 	"fmt"
 	"os"
-
-	"encoding/json"
 
 	"github.com/amazeeio/lagoon-cli/api"
 	"github.com/amazeeio/lagoon-cli/app"
 	"github.com/amazeeio/lagoon-cli/graphql"
+	"github.com/amazeeio/lagoon-cli/lagoon/projects"
 	"github.com/logrusorgru/aurora"
 	"github.com/spf13/cobra"
 )
@@ -42,33 +42,19 @@ var deleteProjectCmd = &cobra.Command{
 		}
 		projectName := args[0]
 
-		lagoonAPI, err := graphql.LagoonAPI()
-		if err != nil {
-			fmt.Println(err)
-			return
-		}
-
-		var jsonBytes []byte
-		project := api.Project{
-			Name: projectName,
-		}
-
 		fmt.Println(fmt.Sprintf("Deleting %s", projectName))
 
 		if yesNo() {
-			projectByName, err := lagoonAPI.DeleteProject(project)
+			deleteResult, err := projects.DeleteProject(projectName)
 			if err != nil {
 				fmt.Println(err)
 				return
 			}
-			jsonBytes, _ = json.Marshal(projectByName)
-			reMappedResult := projectByName.(map[string]interface{})
-			jsonBytes, _ = json.Marshal(reMappedResult["deleteProject"])
 
-			if string(jsonBytes) == "success" {
-				fmt.Println(fmt.Sprintf("Result: %s", aurora.Green(string(jsonBytes))))
+			if string(deleteResult) == "success" {
+				fmt.Println(fmt.Sprintf("Result: %s", aurora.Green(string(deleteResult))))
 			} else {
-				fmt.Println(fmt.Sprintf("Result: %s", aurora.Yellow(string(jsonBytes))))
+				fmt.Println(fmt.Sprintf("Result: %s", aurora.Yellow(string(deleteResult))))
 			}
 		}
 	},
@@ -85,32 +71,13 @@ var addProjectCmd = &cobra.Command{
 		}
 		projectName := args[0]
 
-		var jsonBytes []byte
-
-		lagoonAPI, err := graphql.LagoonAPI()
+		addResult, err := projects.AddProject(projectName, jsonPatch)
 		if err != nil {
 			fmt.Println(err)
 			return
 		}
-		project := api.ProjectPatch{}
-		err = json.Unmarshal([]byte(jsonPatch), &project)
-		if err != nil {
-			fmt.Println(err)
-			return
-		}
-		project.Name = projectName
-
-		projectAddResult, err := lagoonAPI.AddProject(project, graphql.ProjectByNameFragment)
-		if err != nil {
-			fmt.Println(err)
-			return
-		}
-		jsonBytes, _ = json.Marshal(projectAddResult)
-
-		reMappedResult := projectAddResult.(map[string]interface{})
 		var addedProject api.Project
-		jsonBytes, _ = json.Marshal(reMappedResult["addProject"])
-		err = json.Unmarshal([]byte(jsonBytes), &addedProject)
+		err = json.Unmarshal([]byte(addResult), &addedProject)
 
 		if err != nil {
 			fmt.Println(err)
@@ -127,8 +94,46 @@ var addProjectCmd = &cobra.Command{
 	},
 }
 
+var updateProjectCmd = &cobra.Command{
+	Use:   "project [project name]",
+	Short: "Update a project",
+	Run: func(cmd *cobra.Command, args []string) {
+		if len(args) < 1 {
+			fmt.Println("Not enough arguments. Requires: project name")
+			cmd.Help()
+			os.Exit(1)
+		}
+		projectName := args[0]
+
+		projectUpdateID, err := projects.UpdateProject(projectName, jsonPatch)
+		if err != nil {
+			fmt.Println(errorFormat(err.Error(), JSON))
+			return
+		}
+		var updatedProject api.Project
+		err = json.Unmarshal([]byte(projectUpdateID), &updatedProject)
+		if err != nil {
+			fmt.Println(errorFormat(err.Error(), JSON))
+			return
+		}
+
+		fmt.Println(fmt.Sprintf("Result: %s", aurora.Green("success")))
+		fmt.Println(fmt.Sprintf("%s: %s", aurora.Yellow("Project"), updatedProject.Name))
+
+	},
+}
+
 func init() {
 	addCmd.AddCommand(addProjectCmd)
 	addProjectCmd.Flags().StringVarP(&jsonPatch, "json", "j", "", "JSON string to patch")
 	deleteCmd.AddCommand(deleteProjectCmd)
+	updateCmd.AddCommand(updateProjectCmd)
+	updateProjectCmd.Flags().StringVarP(&jsonPatch, "json", "j", "", "JSON string to patch")
+	// use json to patch, maybe re-introduce these better later on @TODO
+	// updateProjectCmd.Flags().StringVarP(&projectGitURL, "giturl", "g", "", "GitURL of the project")
+	// updateProjectCmd.Flags().StringVarP(&projectBranches, "branches", "b", "", "Branches of the project")
+	// updateProjectCmd.Flags().StringVarP(&projectProductionEnvironment, "prod-env", "P", "", "Production environment of the project")
+	// updateProjectCmd.Flags().StringVarP(&projectPullRequests, "pull-requests", "r", "", "Pull requests of the project")
+	// updateProjectCmd.Flags().IntVarP(projectAutoIdle, "auto-idle", "a", 1, "Auto idle setting of the project")
+	// updateProjectCmd.Flags().IntVarP(&projectDevelopmentEnvironmentsLimit, "dev-env-limit", "d", 5, "Auto idle setting of the project")
 }
